weather_station_iot_project/RPi_Core/iot_core/requirements.txt
h5py==2.10.0
numpy==1.17.3
paho_mqtt==1.4.0
